#56 Add confirmation prompt buttons for set and delete model

+++ client/src/components/ModelVersions.svelte
@@ -5,6 +5,8 @@
     import { models, activeModel, type Model, type Hyperparameters } from "../stores/modelStore";
 
     let expandedModelVersion: string | null = null;
+    let isConfirmingActivation: boolean = false;
+    let isConfirmingDeletion: boolean = false;
 
     async function getAllModels(): Promise<void> {
         try {
@@ -43,6 +45,7 @@
         } catch (error) {
             console.error("Setting active model failed:", error);
         }
+        resetConfirmationState()
     }
 
     async function deleteModel(model: Model): Promise<void> {
@@ -60,12 +63,29 @@
                 return currentActiveModel; // keep current activemodel if not deleted
             });
         } catch (error) {
-            console.error("Setting active model failed:", error);
+            console.error("Deleting active model failed:", error);
         }
+        resetConfirmationState()
     }
 
     function toggleExpandModel(model: Model): void {
         expandedModelVersion = expandedModelVersion === model.version ? null : model.version;
+        resetConfirmationState()
+    }
+
+    function toggleConfirmActivation(): void {
+        isConfirmingActivation = !isConfirmingActivation;
+        isConfirmingDeletion = false;
+    }
+
+    function toggleConfirmDeletion(): void {
+        isConfirmingDeletion = !isConfirmingDeletion;
+        isConfirmingActivation = false;
+    }
+
+    function resetConfirmationState(): void {
+        isConfirmingDeletion = false;
+        isConfirmingActivation = false;
     }
 
     function parseHyperparameters(rawHyperparameters: any): Hyperparameters {
@@ -125,7 +145,7 @@
                                 </span>
                             </div>
                             {#if $activeModel && $activeModel.version === model.version}
-                                <span class="text-green-600 text-xs font-regular">(Current)</span>
+                                <span class="text-green-600 text-xs font-regular">ACTIVE</span>
                             {/if}
                         </button>
                         
@@ -199,22 +219,68 @@
                                     </tr>
                                 </tbody>
                             </table>
-                            <!-- Swap model button -->
-                            <button
-                                type="button"
-                                class="mt-4 w-full py-2 px-4 bg-primary text-white rounded-md text-center cursor-pointer hover:bg-secondary"
-                                on:click={() => setActiveModel(model)}
-                            >
-                                Set as Active Model
-                            </button>
+                            <!-- Activate model button -->
+                            {#if expandedModelVersion === $activeModel?.version}
+                                <button
+                                    type="button"
+                                    class="mt-4 w-full py-2 px-4 bg-gray-200 text-tertiary rounded-md text-center cursor-not-allowed"
+                                    disabled
+                                >
+                                    Model Active
+                                </button>
+                            {:else if isConfirmingActivation}
+                                <div class="flex space-x-2 mt-4 w-full">
+                                    <button
+                                        type="button"
+                                        class="flex-1 py-2 px-4 bg-green-300 text-white rounded-md text-center cursor-pointer hover:bg-green-400"
+                                        on:click={() => setActiveModel(model)}
+                                    >
+                                        Confirm Activation
+                                    </button>
+                                    <button
+                                        type="button"
+                                        class="flex-1 py-2 px-4 bg-primary text-white rounded-md text-center cursor-pointer hover:bg-secondary"
+                                        on:click={() => toggleConfirmActivation()}
+                                    >
+                                        Cancel
+                                    </button>
+                                </div>
+                            {:else}
+                                <button
+                                    type="button"
+                                    class="mt-4 w-full py-2 px-4 bg-primary text-white rounded-md text-center cursor-pointer hover:bg-secondary"
+                                    on:click={() => toggleConfirmActivation()}
+                                >
+                                    Set as Active Model
+                                </button>
+                            {/if}
                             <!-- Delete model button -->
-                            <button
-                                type="button"
-                                class="mt-2 w-full py-2 px-4 bg-red-500 text-white rounded-md text-center cursor-pointer hover:bg-red-600"
-                                on:click={() => deleteModel(model)}
-                            >
-                                Delete Model
-                            </button>
+                            {#if isConfirmingDeletion}
+                                <div class="flex space-x-2 mt-2 w-full">
+                                    <button
+                                        type="button"
+                                        class="flex-1 py-2 px-4 bg-red-500 text-white rounded-md text-center cursor-pointer hover:bg-red-600"
+                                        on:click={() => deleteModel(model)}
+                                    >
+                                        Confirm Deletion
+                                    </button>
+                                    <button
+                                        type="button"
+                                        class="flex-1 py-2 px-4 bg-primary text-white rounded-md text-center cursor-pointer hover:bg-secondary"
+                                        on:click={() => toggleConfirmDeletion()}
+                                    >
+                                        Cancel
+                                    </button>
+                                </div>
+                            {:else}
+                                <button
+                                    type="button"
+                                    class="mt-2 w-full py-2 px-4 bg-red-500 text-white rounded-md text-center cursor-pointer hover:bg-red-600"
+                                    on:click={() => toggleConfirmDeletion()}
+                                >
+                                    Delete Model
+                                </button>
+                            {/if}
                         </div>
                         {/if}
                     </div>
